+++ src/index.ts
@@ -4,7 +4,4 @@ export * from './handlers/index.js';
 
 export type { Plugin, Filter, Options, RequestHandler } from './types.js';
 
-/**
- * Default plugins
- */
-export * from './plugins/default/index.js';
+export * from './plugins/index.js';